Update for server utils
