+++ examples/ML+DL-Examples/Spark-DL/dl_inference/dataproc/setup/init_spark_dl.sh
@@ -49,7 +49,7 @@ if [[ "${ROLE}" == 'Master' ]]; then
     if gsutil -q stat gs://${SPARK_DL_HOME}/notebooks/**; then
         mkdir spark-dl-notebooks
         gcloud storage cp -r gs://${SPARK_DL_HOME}/notebooks/* spark-dl-notebooks
-        gcloud storage cp gs://${SPARK_DL_HOME}/pytriton_utils.py .
+        gcloud storage cp gs://${SPARK_DL_HOME}/server_utils.py .
     else
         echo "Failed to retrieve notebooks from gs://${SPARK_DL_HOME}/notebooks/"
         exit 1
